refactor and update adapter params
