+++ stories/react/.storybook/preview.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
 import type { Preview } from '@storybook/react';
-import { SearchBox, Hits } from 'react-instantsearch';
-
-import TrieveInstantsearchProvider from '../src/trieveInstantsearchProvider';
+import Playgrond from '../src/components/Playground';
+import TrieveInstantsearchProvider from '../src/components/TrieveInstantsearchProvider';
 
 import './css/index.css';
 
@@ -22,20 +21,7 @@ const preview: Preview = {
         <TrieveInstantsearchProvider>
           <h1 style={{ marginBottom: 8, color: 'gray' }}>Preview</h1>
           <Story />
-
-          <div
-            style={{
-              marginTop: 24,
-              borderTop: '1px solid #f2f2f2',
-              paddingTop: 8,
-            }}
-          >
-            <h1 style={{ marginBottom: 8, color: 'gray' }}>Playground</h1>
-            <SearchBox />
-            <div style={{ marginTop: 8 }}>
-              <Hits />
-            </div>
-          </div>
+          <Playgrond />
         </TrieveInstantsearchProvider>
       );
     },
